Reformat connector, added new stream still does not stop
