+++ airbyte-integrations/connectors/source-bitrix24-crm/integration_tests/configured_catalog.json
@@ -2,21 +2,12 @@
   "streams": [
     {
       "stream": {
-        "name": "customers",
+        "name": "lists_element",
         "json_schema": {},
         "supported_sync_modes": ["full_refresh"]
       },
       "sync_mode": "full_refresh",
       "destination_sync_mode": "overwrite"
-    },
-    {
-      "stream": {
-        "name": "employees",
-        "json_schema": {},
-        "supported_sync_modes": ["full_refresh", "incremental"]
-      },
-      "sync_mode": "incremental",
-      "destination_sync_mode": "append"
     }
   ]
 }
