+++ airbyte-integrations/connectors/source-bitrix24-crm/source_bitrix24_crm/schemas/lists_element.json
