+++ airbyte-integrations/connectors/source-bitrix24-crm/source_bitrix24_crm/streams/deals.py
@@ -1,0 +1,6 @@
+from .base import ObjectListStream
+
+
+class Deals(ObjectListStream):
+    def path(self, **kwargs) -> str:
+        return "crm.deal.list"
